scripts/requirements_convert.txt
openvino-dev[pytorch]
transformers
sentencepiece
onnx
onnxsim